backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic-settings==2.1.0

# AI and Agent Libraries
openai==1.3.7
pydantic==2.5.0

# HTTP and Async
httpx==0.25.2
aiofiles==23.2.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Additional Tools
requests==2.31.0
beautifulsoup4==4.12.2
